集成mybatis-plus，修改旧有代码

+++ easy-iot-server/src/main/java/com/itheimaxia/easyiot/server/bean/ConnectionDO.java
@@ -17,7 +17,7 @@ public class ConnectionDO implements Serializable {
     private String ip_address;
     private Integer proto_ver;
     private Integer connected_at;
-    private Long disconnect_at;
+    private Integer disconnect_at;
     @Id
     private String username;
